+++ src/main/java/gob/gt/infom/repositories/AmortizacionRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 import gob.gt.infom.models.Amortizacion;
 
 
-public interface AmortizacionRepository extends CrudRepository<Amortizacion, Long> {
+public interface AmortizacionRepository extends CrudRepository<Amortizacion, Integer> {
 
-  Iterable<Amortizacion> findAllByPrestamoId(Long id_prestamo);
+  Iterable<Amortizacion> findAllByPrestamoId(Integer id_prestamo);
 
 }
