+++ src/main/java/gob/gt/infom/repositories/EstadoCivilRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gob.gt.infom.models.EstadoCivil;
 
-public interface EstadoCivilRepository extends CrudRepository<EstadoCivil, Long> {
+public interface EstadoCivilRepository extends CrudRepository<EstadoCivil, Integer> {
   
 }
