+++ src/main/java/gob/gt/infom/repositories/FuncionarioRepository.java
@@ -3,5 +3,5 @@ package gob.gt.infom.repositories;
 import gob.gt.infom.models.Funcionario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FuncionarioRepository extends CrudRepository<Funcionario, Long> {
+public interface FuncionarioRepository extends CrudRepository<Funcionario, Integer> {
 }
