+++ src/main/java/gob/gt/infom/repositories/MenuRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import gob.gt.infom.models.Menu;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuRepository extends CrudRepository<Menu, Long> {
+public interface MenuRepository extends CrudRepository<Menu, Integer> {
 
   public List<Menu> findAllByOrderByOrdenAsc();
