+++ src/main/java/gob/gt/infom/repositories/PuestoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gob.gt.infom.models.Puesto;
 
-public interface PuestoRepository extends CrudRepository<Puesto, Long> {
+public interface PuestoRepository extends CrudRepository<Puesto, Integer> {
   
 }
