+++ src/main/java/gob/gt/infom/repositories/RegionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gob.gt.infom.models.Region;
 
-public interface RegionRepository extends CrudRepository<Region, Long> {
+public interface RegionRepository extends CrudRepository<Region, Integer> {
   
 }
