+++ src/main/java/gob/gt/infom/repositories/RegionalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gob.gt.infom.models.Regional;
 
-public interface RegionalRepository extends CrudRepository<Regional, Long> {
+public interface RegionalRepository extends CrudRepository<Regional, Integer> {
   
 }
