+++ src/main/java/gob/gt/infom/repositories/RolRepository.java
@@ -3,5 +3,5 @@ package gob.gt.infom.repositories;
 import gob.gt.infom.models.Rol;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RolRepository extends CrudRepository<Rol, Long> {
+public interface RolRepository extends CrudRepository<Rol, Integer> {
 }
