+++ src/main/java/gob/gt/infom/repositories/SubmenuRepository.java
@@ -4,5 +4,5 @@ import gob.gt.infom.models.Submenu;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubmenuRepository extends CrudRepository<Submenu, Long> {
+public interface SubmenuRepository extends CrudRepository<Submenu, Integer> {
 }
